python_web_framework/petstagram/requirements.txt
antiorm==1.2.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
django-bootstrap4==2.3.1
django-forms-bootstrap==3.1.0
entrypoints==0.3
filelock==3.0.12
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.7
jupyterlab-server==1.2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
virtualenv==20.0.35
wcwidth==0.2.5
webencodings==0.5.1
